jpms uses new config and rename http2 example to https
