+++ jpms-examples/src/main/java/io/vertx/example/jpms/https/Server.java
@@ -1,10 +1,11 @@
-package io.vertx.example.jpms.http2;
+package io.vertx.example.jpms.https;
 
 import io.vertx.core.*;
 import io.vertx.core.http.HttpServer;
-import io.vertx.core.http.HttpServerOptions;
+import io.vertx.core.http.HttpServerConfig;
 import io.vertx.core.json.JsonObject;
 import io.vertx.core.net.JksOptions;
+import io.vertx.core.net.ServerSSLOptions;
 
 public class Server extends VerticleBase {
 
@@ -16,13 +17,10 @@ public class Server extends VerticleBase {
 
   @Override
   public Future<?> start() {
-    HttpServerOptions options = new HttpServerOptions()
-      .setUseAlpn(true)
-      .setKeyCertOptions(new JksOptions().setPath("server-keystore.jks").setPassword("wibble"))
-      .setSsl(true);
-
     HttpServer server = vertx
-      .createHttpServer(options)
+      .createHttpServer(
+        new HttpServerConfig().setSsl(true),
+        new ServerSSLOptions().setKeyCertOptions(new JksOptions().setPath("server-keystore.jks").setPassword("wibble")))
       .requestHandler(req -> {
         req.response().end(new JsonObject()
           .put("http", req.version())
