+++ jpms-examples/src/main/java/io/vertx/example/jpms/openssl/Server.java
@@ -2,10 +2,11 @@ package io.vertx.example.jpms.openssl;
 
 import io.vertx.core.*;
 import io.vertx.core.http.HttpServer;
-import io.vertx.core.http.HttpServerOptions;
+import io.vertx.core.http.HttpServerConfig;
 import io.vertx.core.json.JsonObject;
 import io.vertx.core.net.JksOptions;
 import io.vertx.core.net.OpenSSLEngineOptions;
+import io.vertx.core.net.ServerSSLOptions;
 
 public class Server extends VerticleBase {
 
@@ -17,15 +18,13 @@ public class Server extends VerticleBase {
 
   @Override
   public Future<?> start() {
-    HttpServerOptions options = new HttpServerOptions()
-      .setSslEngineOptions(new OpenSSLEngineOptions())
-      .setKeyCertOptions(new JksOptions()
+    HttpServer server = vertx.httpServerBuilder()
+      .with(new HttpServerConfig().setSsl(true))
+      .with(new OpenSSLEngineOptions())
+      .with(new ServerSSLOptions().setKeyCertOptions(new JksOptions()
         .setPath("server-keystore.jks")
-        .setPassword("wibble"))
-      .setSsl(true);
-
-    HttpServer server = vertx
-      .createHttpServer(options)
+        .setPassword("wibble")))
+      .build()
       .requestHandler(req -> {
         req.response().end(new JsonObject()
           .put("http", req.version())
